feat: Implement PlayerStatsScreen with scrollable layout
- Add PlayerStatsScreen with character stats display
- Extract UI components into separate composables
- Implement vertical scrolling for landscape mode compatibility
- Add character progression and level-up functionality
