+++ app/src/main/java/com/github/cstrerath/uncover/ui/activities/PlayerStatsActivity.kt
@@ -3,7 +3,7 @@ package com.github.cstrerath.uncover.ui.activities
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import com.github.cstrerath.uncover.ui.base.BaseActivity
-import com.github.cstrerath.uncover.ui.screens.PlayerStatsScreen
+import com.github.cstrerath.uncover.ui.screens.playerDetails.PlayerStatsScreen
 
 
 class PlayerStatsActivity : BaseActivity() {
