+++ app/src/main/java/com/github/cstrerath/uncover/ui/screens/playerDetails/ExperienceBar.kt
@@ -1,0 +1,52 @@
+package com.github.cstrerath.uncover.ui.screens.playerDetails
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.LinearProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.github.cstrerath.uncover.R
+
+@Composable
+internal fun ExperienceBar(experience: Int, requiredXp: Int, level: Int) {
+    Column {
+        if (level >= 25) {
+            Text(
+                text = stringResource(R.string.max_level_reached),
+                style = MaterialTheme.typography.titleSmall,
+                color = MaterialTheme.colorScheme.primary
+            )
+            LinearProgressIndicator(
+                progress = { 1f },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(8.dp)
+                    .clip(RoundedCornerShape(4.dp)),
+                color = MaterialTheme.colorScheme.primary,
+                trackColor = MaterialTheme.colorScheme.surfaceVariant,
+            )
+        } else {
+            val displayXp = experience.coerceAtMost(requiredXp)
+            Text(
+                text = "Experience: $displayXp / $requiredXp",
+                style = MaterialTheme.typography.titleSmall
+            )
+            LinearProgressIndicator(
+                progress = { (displayXp.toFloat() / requiredXp).coerceIn(0f, 1f) },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(8.dp)
+                    .clip(RoundedCornerShape(4.dp)),
+                color = MaterialTheme.colorScheme.secondary,
+                trackColor = MaterialTheme.colorScheme.surfaceVariant,
+            )
+        }
+    }
+}
