+++ app/src/main/java/com/github/cstrerath/uncover/ui/screens/playerDetails/LevelDisplay.kt
@@ -1,0 +1,20 @@
+package com.github.cstrerath.uncover.ui.screens.playerDetails
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+
+@Composable
+internal fun LevelDisplay(level: Int) {
+    Text(
+        text = "Level $level",
+        style = MaterialTheme.typography.headlineSmall,
+        fontWeight = FontWeight.Bold,
+        modifier = Modifier.padding(vertical = 8.dp),
+        color = MaterialTheme.colorScheme.primary
+    )
+}
