+++ app/src/main/java/com/github/cstrerath/uncover/ui/screens/playerDetails/LevelUpDialog.kt
@@ -1,0 +1,25 @@
+package com.github.cstrerath.uncover.ui.screens.playerDetails
+
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.github.cstrerath.uncover.R
+
+@Composable
+internal fun LevelUpDialog(
+    message: String,
+    onDismiss: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(text = stringResource(R.string.level_up_information)) },
+        text = { Text(message) },
+        confirmButton = {
+            Button(onClick = onDismiss) {
+                Text("OK")
+            }
+        }
+    )
+}
